Implemented "same" filter
That took a lot of debugging over the past week. Turns out, the reason
it was giving infinite loops (this.generateSpriteCommandSameFromRender
<--> this.decode) was that BaseFiler.get(key) was always returning the
Render object even if it was replaced in its container with the new
directory.

BaseFiler now calls clearCached on the key. There was actually a glitch
in StringFilr where clearCached was only deleting the normal-replaced
version of the key, not just the key.